+++ Changelog.md
@@ -1,6 +1,6 @@
 # Fossa CLI Changelog
 
-## Unreleased
+## v3.0.17
 
 - Npm: Fixes an issue where a package-lock.json dep with a boolean 'resolved' key wouldn't parse. ([#775](https://github.com/fossas/fossa-cli/pull/775))
 - Npm: Fixes an issue where analyzing `package-lock.json` would miss duplicate packages with different versions. ([#779](https://github.com/fossas/fossa-cli/pull/779))